Update README.md
The way I will be using the symbols might not be correct. Remember to use your imagination and to look at it in all directions. We can eventually have our own line, forever expanding out. We have to remember to keep in contact connections like a web or we become weak.